requirements.txt
aiogram==3.15.0
httpx==0.28.1
redis==5.2.1
python-dotenv==1.0.1
